refactor command center exception resolution router
